Initial changes to new format
