+++ src/frontend/src/lib/stores/user-control-store.ts
@@ -1,0 +1,4 @@
+import type { UserId } from "$lib/types/user";
+import { initCreatedStore } from "./user-created-store";
+
+export const userIdCreatedStore = initCreatedStore<UserId | string>();
