+++ src/frontend/src/lib/types/store.ts
@@ -1,0 +1,4 @@
+export interface CreatedData<T> {
+    data: T;
+    certified: boolean;
+}
