+++ src/frontend/src/lib/types/user.ts
@@ -1,0 +1,3 @@
+import type { Principal } from "@dfinity/principal";
+
+export type UserId = Principal;
