+++ src/frontend/src/lib/types/utils.ts
@@ -1,0 +1,3 @@
+// We disable the eslint rule here because this is the utility type that we use to define such rule.
+// eslint-disable-next-line local-rules/use-option-type-wrapper
+export type Option<T> = T | null | undefined;
