+++ src/frontend/src/routes/(private)/+layout.ts
@@ -1,0 +1,1 @@
+export const ssr = false;
