+++ src/frontend/src/routes/(public)/+layout.svelte
@@ -1,0 +1,10 @@
+<script lang="ts">
+    import type { Snippet } from 'svelte';
+
+    interface Props {
+        children: Snippet;
+    }
+    let { children }: Props = $props();
+</script>
+
+{@render children()}
