+++ src/frontend/src/routes/(public)/+layout.ts
@@ -1,0 +1,1 @@
+export const trailingSlash = "always";
