--- src/frontend/src/routes/governance/+page.svelte
@@ -1,9 +1,0 @@
-<script lang="ts">
-    import ProposalsList from "$lib/components/goverance/proposals-list.svelte";
-    import Layout from "../Layout.svelte";
-
-</script>
-<Layout>
-    <ProposalsList />
-</Layout>
-
